Migrate `Ui` to `ConsoleUi`
`ConsoleUi` now inherits from abstract class `Ui`. This allows creation
of a new class with JavaFX interaction inheriting from `Ui`, allowing
both GUI/text-mode interfaces for Duke.
