+++ src/main/java/duke/command/AddEventCommand.java
@@ -2,7 +2,7 @@ package duke.command;
 
 import duke.Storage;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 import duke.task.Event;
 import duke.task.Task;
