+++ src/main/java/duke/command/ByeCommand.java
@@ -2,7 +2,7 @@ package duke.command;
 
 import duke.Storage;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 
 public class ByeCommand extends Command {
     /**
