+++ src/main/java/duke/command/CompleteTaskCommand.java
@@ -2,7 +2,7 @@ package duke.command;
 
 import duke.Storage;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 import duke.task.Task;
 
 public class CompleteTaskCommand extends Command {
